requirements.txt
python-nmap
ipaddress
colorama
paramiko
pycurl
ajpy
pyopenssl
mysqlclient
psycopg2
pycrypto
dnspython
IPy
pysnmp
pyasn1
Flask-SQLAlchemy